pyproject.toml
[project]
name = "github-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "matplotlib>=3.9.3",
    "pandas>=2.2.3",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.10.2",
    "python-dateutil>=2.9.0.post0",
    "sqlmodel>=0.0.22",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "mypy>=1.13.0",
    "python-lsp-server[all]>=1.12.0",
    "ruff>=0.8.1",
]
